rename domainhooks.rb to domain_hooks.rb
