+++ common/lib/symbiosis/domain_hooks.rb
